requirements.txt
requests
colorlog
python-dotenv
